Perform multiple searches before ending session
