+++ MoochKick/MoochKick/MoochKick.csproj
@@ -59,7 +59,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Monolith.cs" />
+    <Compile Include="Program.cs" />
     <Compile Include="Player.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SpartanCompany.cs" />
